// src/api/models/work-experience-im.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Resume Builder API
 * An API for Online Resume Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface WorkExperienceIM
 */
export interface WorkExperienceIM {
    /**
     * 
     * @type {string}
     * @memberof WorkExperienceIM
     */
    resumeId: string;
    /**
     * 
     * @type {string}
     * @memberof WorkExperienceIM
     */
    companyName: string;
    /**
     * 
     * @type {string}
     * @memberof WorkExperienceIM
     */
    position: string;
    /**
     * 
     * @type {string}
     * @memberof WorkExperienceIM
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof WorkExperienceIM
     */
    startDate: string;
    /**
     * 
     * @type {string}
     * @memberof WorkExperienceIM
     */
    endDate?: string | null;
}