// src/api/models/work-experience-vm.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Resume Builder API
 * An API for Online Resume Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DateOnly } from './date-only';
/**
 * 
 * @export
 * @interface WorkExperienceVM
 */
export interface WorkExperienceVM {
    /**
     * 
     * @type {string}
     * @memberof WorkExperienceVM
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof WorkExperienceVM
     */
    resumeId: string;
    /**
     * 
     * @type {string}
     * @memberof WorkExperienceVM
     */
    companyName: string;
    /**
     * 
     * @type {string}
     * @memberof WorkExperienceVM
     */
    position: string;
    /**
     * 
     * @type {string}
     * @memberof WorkExperienceVM
     */
    description: string;
    /**
     * 
     * @type {DateOnly}
     * @memberof WorkExperienceVM
     */
    startDate: DateOnly;
    /**
     * 
     * @type {DateOnly}
     * @memberof WorkExperienceVM
     */
    endDate?: DateOnly;
}